% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_reload.R
\name{bg_gcs_auth}
\alias{bg_gcs_auth}
\title{Authenticate bgScorer to GCS}
\usage{
bg_gcs_auth(
  file_name = "bgScorer-testing.json",
  use_gcs = get_golem_config("use_gcs")
)
}
\arguments{
\item{file_name}{Name of file containing Service Account key}

\item{use_gcs}{Whether to use GCS. Default is the "use_gcs" config value}
}
\value{
Function to be called when starting the app
}
\description{
Authenticate bgScorer to GCS
}
