% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgScorer-package.R
\docType{package}
\name{bgScorer-package}
\alias{bgScorer}
\alias{bgScorer-package}
\title{bgScorer: Betting Game Scorer}
\description{
Shiny application to keep score in the Betting Game (AKA Contract Whist).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jaron94/Betting-Game-Scorer}
  \item Report bugs at \url{https://github.com/jaron94/Betting-Game-Scorer/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Aron \email{jonathan.aron94@gmail.com}

}
\keyword{internal}
