% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_reload.R
\name{google_auth_config}
\alias{google_auth_config}
\title{Configure Google Authentication using service account}
\usage{
google_auth_config(
  base = Sys.getenv("GOOGLE_CREDS_BASE"),
  complete = Sys.getenv("GOOGLE_APPLICATION_CREDENTIALS"),
  pk = Sys.getenv("GOOGLE_PK")
)
}
\arguments{
\item{base}{Path to base for service account token}

\item{complete}{Path to completed token}

\item{pk}{Private key}
}
\description{
Configure Google Authentication using service account
}
