% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_reload.R
\name{googlesheets4_partials}
\alias{googlesheets4_partials}
\alias{bg_write_sheet}
\alias{bg_read_sheet}
\alias{bg_sheet_append}
\alias{bg_sheet_names}
\alias{bg_sheet_delete}
\title{Use googlesheets4}
\description{
Mapping from bg_* to googlesheets4 functions
}
\details{
\itemize{
\item \link{bg_write_sheet} -- \link[googlesheets4:sheet_write]{googlesheets4::sheet_write}
\item \link{bg_read_sheet} -- \link[googlesheets4:range_read]{googlesheets4::read_sheet}
\item \link{bg_sheet_append} -- \link[googlesheets4:sheet_append]{googlesheets4::sheet_append}
\item \link{bg_sheet_names} -- \link[googlesheets4:sheet_delete]{googlesheets4::sheet_delete}
\item \link{bg_sheet_delete} -- \link[googlesheets4:sheet_delete]{googlesheets4::sheet_delete}
}
}
