% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_reload.R
\name{save_game}
\alias{save_game}
\title{Save game data to a Google sheet}
\usage{
save_game(game_id, data, in_progress)
}
\arguments{
\item{game_id}{String containing ID of the game}

\item{data}{Data frame containing game data}

\item{in_progress}{TRUE if game is in progress, FALSE if game is completed}
}
\value{
Invisibly, the name of the sheet where the data was saved
}
\description{
Save game data to a Google sheet
}
