% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_reload.R
\name{bg_gcs_auth}
\alias{bg_gcs_auth}
\title{Authenticate bgScorer to GCS}
\usage{
bg_gcs_auth(json_file = Sys.getenv("GOOGLE_APPLICATION_CREDENTIALS"))
}
\value{
Function to be called when starting the app
}
\description{
Authenticate bgScorer to GCS
}
